+++ src/shared/lib/api/user.ts
@@ -1,0 +1,11 @@
+export type IUserName = {
+  id: number;
+  username: string;
+  fullName: string;
+};
+
+export type IUserSummary = {
+  id: number;
+  username: string;
+  image: string;
+};
